+++ backend/src/main/java/hyfive/gachita/dispatch/UnDispatchService.java
@@ -1,14 +1,9 @@
 package hyfive.gachita.dispatch;
 
-import hyfive.gachita.dispatch.dto.DispatchLocation;
 import hyfive.gachita.dispatch.module.filter.BoundingBoxFilter;
 import hyfive.gachita.dispatch.module.filter.HaversineFilter;
-import hyfive.gachita.dispatch.module.filter.condition.BoundingBoxCondition;
-import hyfive.gachita.dispatch.module.filter.condition.RadiusCondition;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
