+++ backend/src/main/java/hyfive/gachita/dispatch/module/condition/PathCondition.java
@@ -1,4 +1,4 @@
-package hyfive.gachita.dispatch.module.filter.condition;
+package hyfive.gachita.dispatch.module.condition;
 
 import java.time.LocalTime;
 
@@ -6,5 +6,5 @@ public record PathCondition (
         LocalTime maybeOnTime,
         LocalTime deadline,
         boolean walker
-) implements FilterCondition {
+) implements Condition {
 }
