+++ backend/src/main/java/hyfive/gachita/dispatch/module/filter/BoundingBoxFilter.java
@@ -1,7 +1,7 @@
 package hyfive.gachita.dispatch.module.filter;
 
 import hyfive.gachita.dispatch.dto.FilterDto;
-import hyfive.gachita.dispatch.module.filter.condition.BoundingBoxCondition;
+import hyfive.gachita.dispatch.module.condition.BoundingBoxCondition;
 import org.springframework.stereotype.Component;
 import java.util.List;
 import java.util.stream.Collectors;
