+++ backend/src/main/java/hyfive/gachita/dispatch/module/provider/DispatchLocationProvider.java
@@ -1,0 +1,22 @@
+package hyfive.gachita.dispatch.module.provider;
+
+import hyfive.gachita.dispatch.dto.DispatchLocation;
+import hyfive.gachita.dispatch.module.condition.Condition;
+
+import java.util.List;
+
+/**
+ * 요약 : 배차 시스템 dto로 변경해주는 Provider
+ * 설명 : 배차 시스템 작동을 위해 필요한 정보를 DispatchLocation 타입으로 변경해줍니다
+ * 사용 방법 :
+ *  - DispatchLocationProvider interface를 구현
+ */
+public interface DispatchLocationProvider {
+    default List<DispatchLocation> getAll() {
+        throw new UnsupportedOperationException("getAll() not supported");
+    }
+
+    default List<DispatchLocation> getByCondition(Condition condition) {
+        throw new UnsupportedOperationException("getByCondition() not supported");
+    }
+}
